Convert `Box/index.js`
