+++ packages/fuselage/src/types/Falsy.ts
@@ -1,0 +1,1 @@
+export type Falsy = false | 0 | '' | null | undefined;
